:truck: chore(utils): move base resource service from `services` to `utils`
