+++ src/services/resources/categories.ts
@@ -1,4 +1,4 @@
-import BaseResourceService from './base-resource-service'
+import BaseResourceService from '../../utils/services/base-resource-service'
 
 class CategoriesService extends BaseResourceService {
   // add custom methods and properties if needed
